requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.27.2=pypi_0
aiofiles=23.2.1=pypi_0
aiohappyeyeballs=2.4.4=py310h06a4308_0
aiohttp=3.11.10=py310h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
arrow-cpp=19.0.0=h865e1df_2
async-timeout=5.0.1=py310h06a4308_0
attrs=25.3.0=pypi_0
aws-c-auth=0.6.19=h5eee18b_0
aws-c-cal=0.5.20=hdbd6064_0
aws-c-common=0.8.5=h5eee18b_0
aws-c-compression=0.2.16=h5eee18b_0
aws-c-event-stream=0.2.15=h6a678d5_0
aws-c-http=0.6.25=h5eee18b_0
aws-c-io=0.13.10=h5eee18b_0
aws-c-mqtt=0.7.13=h5eee18b_0
aws-c-s3=0.1.51=hdbd6064_0
aws-c-sdkutils=0.1.6=h5eee18b_0
aws-checksums=0.1.13=h5eee18b_0
aws-crt-cpp=0.18.16=h6a678d5_0
aws-sdk-cpp=1.11.212=hecad206_0
bitsandbytes=0.41.1=pypi_0
boost-cpp=1.82.0=hdb19cb5_2
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.4.26=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.2.1=pypi_0
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
datasets=3.3.2=py310h06a4308_0
deepspeed=0.12.6=pypi_0
dill=0.3.4=pyhd3eb1b0_0
dnspython=2.7.0=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.6.1=pypi_0
einops-exts=0.0.4=pypi_0
email-validator=2.2.0=pypi_0
exceptiongroup=1.3.0=pypi_0
fastapi=0.111.0=pypi_0
fastapi-cli=0.0.7=pypi_0
ffmpy=0.5.0=pypi_0
filelock=3.18.0=pypi_0
fonttools=4.58.1=pypi_0
frozenlist=1.5.0=py310h5eee18b_0
fsspec=2024.12.0=py310h06a4308_0
gflags=2.2.2=h6a678d5_1
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
glog=0.5.0=h6a678d5_1
gradio=4.16.0=pypi_0
gradio-client=0.8.1=pypi_0
h11=0.14.0=pypi_0
hf-xet=1.1.2=pypi_0
hjson=3.1.0=pypi_0
httpcore=0.17.3=pypi_0
httptools=0.6.4=pypi_0
httpx=0.24.0=pypi_0
huggingface-hub=0.32.3=pypi_0
huggingface_hub=0.31.2=py310h06a4308_0
icu=73.1=h6a678d5_0
idna=3.10=pypi_0
importlib-resources=6.5.2=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.5.1=pypi_0
jsonschema=4.24.0=pypi_0
jsonschema-specifications=2025.4.1=pypi_0
kiwisolver=1.4.8=pypi_0
krb5=1.20.1=h143b758_1
latex2mathml=3.78.0=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libabseil=20250127.0=cxx17_h6a678d5_0
libboost=1.82.0=h109eef0_2
libbrotlicommon=1.0.9=h5eee18b_9
libbrotlidec=1.0.9=h5eee18b_9
libbrotlienc=1.0.9=h5eee18b_9
libcurl=8.12.1=hc9e6f67_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libgrpc=1.71.0=h2d74bed_0
libnghttp2=1.57.0=h2d74bed_0
libprotobuf=5.29.3=hc99497a_0
libre2-11=2024.07.02=h6a678d5_0
libssh2=1.11.1=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libthrift=0.15.0=h1795dd8_2
libuuid=1.41.5=h5eee18b_0
llava=1.2.2.post1=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown-it-py=3.0.0=pypi_0
markdown2=2.5.3=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.3=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=py310h5eee18b_0
multiprocess=0.70.12.2=py310h7f8727e_0
narwhals=1.41.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.4.2=pypi_0
ninja=1.11.1.4=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-ml-py=12.575.51=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.9.41=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
openssl=3.0.16=h5eee18b_0
orc=2.1.1=hd396ef6_0
orjson=3.10.18=pypi_0
packaging=25.0=pypi_0
pandas=2.2.3=pypi_0
peft=0.10.0=pypi_0
pillow=10.4.0=pypi_0
pip=25.1.1=pypi_0
platformdirs=4.3.8=pypi_0
propcache=0.3.1=py310h5eee18b_0
protobuf=6.31.1=pypi_0
psutil=7.0.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=19.0.0=py310h6a678d5_1
pydantic=2.11.5=pypi_0
pydantic-core=2.33.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.19.1=pypi_0
pynvml=12.0.0=pypi_0
pyparsing=3.2.3=pypi_0
python=3.10.16=he870216_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.1.0=pypi_0
python-multipart=0.0.20=pypi_0
python-xxhash=3.5.0=py310h5eee18b_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
re2=2024.07.02=hdb19cb5_0
readline=8.2=h5eee18b_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=14.0.0=pypi_0
rich-toolkit=0.14.7=pypi_0
rpds-py=0.25.1=pypi_0
ruff=0.11.12=pypi_0
s2n=1.3.27=hdbd6064_0
safetensors=0.5.3=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.15.3=pypi_0
semantic-version=2.10.0=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=2.29.1=pypi_0
setproctitle=1.3.6=pypi_0
setuptools=78.1.1=py310h06a4308_0
shellingham=1.5.4=pypi_0
shortuuid=1.0.13=pypi_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
snappy=1.2.1=h6a678d5_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h5eee18b_0
starlette=0.37.2=pypi_0
svgwrite=1.4.3=pypi_0
sympy=1.14.0=pypi_0
threadpoolctl=3.6.0=pypi_0
timm=0.6.13=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.15.2=pypi_0
tomlkit=0.12.0=pypi_0
torch=2.1.2=pypi_0
torchvision=0.16.2=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.37.2=pypi_0
triton=2.1.0=pypi_0
typer=0.16.0=pypi_0
typing-extensions=4.13.2=pypi_0
typing-inspection=0.4.1=pypi_0
tzdata=2025.2=pypi_0
ujson=5.10.0=pypi_0
urllib3=2.4.0=pypi_0
utf8proc=2.6.1=h5eee18b_1
uvicorn=0.34.3=pypi_0
uvloop=0.21.0=pypi_0
wandb=0.19.11=pypi_0
watchfiles=1.0.5=pypi_0
wavedrom=2.0.3.post3=pypi_0
websockets=11.0.3=pypi_0
wheel=0.45.1=py310h06a4308_0
xxhash=0.8.0=h7f8727e_3
xz=5.6.4=h5eee18b_1
yarl=1.18.0=py310h5eee18b_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
